Project/Back/requirements.txt
# FastAPI Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0

# HTTP 客户端
requests==2.31.0
httpx==0.26.0

# 环境变量管理
python-dotenv==1.0.0

# 数据验证
pydantic==2.5.3

# JSON 处理
orjson==3.9.10

# 异步支持
aiofiles==23.2.1

# ==================== 深度学习依赖（手语识别） ====================
# PyTorch（根据 CUDA 版本选择合适的版本）
torch==2.1.0
torchvision==0.16.0

# 数据处理
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2

# 图像处理
Pillow==10.2.0
opencv-python==4.8.0

# 数据集
datasets==2.14.0

# 进度条
tqdm==4.66.1

# FastAPI 文件上传
python-multipart==0.0.9

# ==================== 可视化依赖（论文图表） ====================
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.3.2
